refactor(loyalty): split AdminLoyalty (526→124) em hook + componentes
Extrai verbatim de src/pages/AdminLoyalty.tsx:
- useAdminLoyalty: state, effects (redirect non-staff + load), loadData
  (3 queries + agregação por usuário), ajuste de pontos e derivados
  econômicos (passivo/taxa de resgate/top recompensas/top saldos)
- AdjustDialog, CustomerDetail, LoyaltyStats, EconomicInsights, CustomerList,
  types.ts, config.ts (TIERS/getTier)
- +17 testes (helpers puros + presentacionais)

Comportamento preservado 1:1. tsc baseline=0, eslint=0.
